+++ packages/core/src/devices/camera/export.ts
@@ -1,0 +1,35 @@
+/**
+ * @file packages/core/src/devices/camera/export.ts
+ * @description Exports the Chapter 16 Camera device modules.
+ * @author Luca Liguori
+ * @created 2026-09-13
+ * @version 1.0.0
+ * @license Apache-2.0
+ *
+ * Copyright 2026, 2027, 2028 Luca Liguori.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* oxlint-disable oxc/no-barrel-file */
+
+// Chapter 16. Camera Device Types - Single class device types
+export * from '../audioDoorbell.js';
+export * from '../camera.js';
+export * from '../cameraController.js';
+export * from '../chime.js';
+export * from '../doorbell.js';
+export * from '../floodlightCamera.js';
+export * from '../intercom.js';
+export * from '../snapshotCamera.js';
+export * from '../videoDoorbell.js';